requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.4.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
segmentation-models-pytorch>=0.3.3

